+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 ## [WIP] 0.6.1-0.1.1 / 2019-January-??
 
 - Upgrade Calfpath to version 0.7.1
-- [Todo] Always emit verbose message in `gossamer.core.inducer/calfpath-routes->ring-handler`
+- Always emit verbose message in `gossamer.core.inducer/calfpath-routes->ring-handler`
 
 
 ## 0.6.1-0.1.0 / 2018-October-10